refactor: add enum to make classes with same name possible
