+++ Client/Client/Api/Two/AccountsTwo.swift
@@ -1,0 +1,16 @@
+// 
+// This file is part of Ark Swift Client.
+//
+// (c) Ark Ecosystem <info@ark.io>
+//
+// For the full copyright and license information, please view the LICENSE
+// file that was distributed with this source code.
+//
+
+import Foundation
+
+extension Two {
+    class Accounts {
+        
+    }
+}
