+++ Client/Client/Api/Two/Two.swift
@@ -1,0 +1,14 @@
+// 
+// This file is part of Ark Swift Client.
+//
+// (c) Ark Ecosystem <info@ark.io>
+//
+// For the full copyright and license information, please view the LICENSE
+// file that was distributed with this source code.
+//
+
+import Foundation
+
+enum Two {
+    
+}
